services/content-service/requirements.txt
fastapi
uvicorn
minio
pymongo
motor
python-multipart
pydantic
python-jose[cryptography]
httpx
prometheus-client
prometheus-fastapi-instrumentator
aio-pika
